+++ sema4ai-python-ls-core/tests/sema4ai_ls_core_tests/test_basic.py
@@ -1,3 +1,6 @@
+import pytest
+
+
 def test_isinstance_name():
     from sema4ai_ls_core.basic import isinstance_name
 
@@ -22,9 +25,10 @@ def test_isinstance_name():
 
 
 def test_notify_about_import(tmpdir):
-    from sema4ai_ls_core.basic import notify_about_import
-    import sys
     import io
+    import sys
+
+    from sema4ai_ls_core.basic import notify_about_import
     from sema4ai_ls_core.core_log import configure_logger
 
     tmpdir.join("my_test_notify_about_import.py").write_text("a = 10", "utf-8")
@@ -46,3 +50,62 @@ def test_notify_about_import(tmpdir):
         )
     finally:
         sys.path = path
+
+
+@pytest.mark.parametrize(
+    "scenario",
+    [
+        "monitor",
+        "future",
+        "timeout",
+    ],
+)
+def test_launch_and_return_future(scenario):
+    import sys
+    from concurrent.futures import CancelledError, TimeoutError
+
+    from sema4ai_ls_core.basic import ProcessResultStatus, launch_and_return_future
+    from sema4ai_ls_core.jsonrpc.monitor import Monitor
+
+    code = """
+import time
+while True:
+    time.sleep(1)
+"""
+    monitor = Monitor()
+    future = launch_and_return_future(
+        [sys.executable, "-c", code],
+        environ={},
+        cwd=".",
+        monitor=monitor,
+        timeout=1 if scenario == "timeout" else 30,
+    )
+
+    with pytest.raises(TimeoutError):
+        future.result(1)
+
+    if scenario == "monitor":
+        monitor.cancel()
+    elif scenario == "future":
+        future.cancel()
+    elif scenario == "timeout":
+        pass
+    else:
+        raise ValueError(f"Invalid scenario: {scenario}")
+
+    if scenario == "monitor":
+        result = future.result(1)
+        assert result.returncode != 0
+        assert result.status == ProcessResultStatus.CANCELLED
+    elif scenario == "future":
+        with pytest.raises(CancelledError):
+            future.result(1)
+    elif scenario == "timeout":
+        result = future.result()
+        assert result.returncode != 0
+        assert result.status == ProcessResultStatus.TIMED_OUT
+    else:
+        raise ValueError(f"Invalid scenario: {scenario}")
+
+    process = future._process_weak_ref()
+    assert process.poll() is not None